requirements.txt
praw==7.7.0
Mastodon.py==1.8.1
transformers==4.42.0
torch==2.2.0
python-dotenv==1.0.0